Image upload, integrated with Summernote, file upload etc...

+++ routes/photos.js
@@ -5,6 +5,8 @@ var PhotoMods = require("../models/photos");
 var multer  = require('multer');
 var path = require('path');
 var mkdirp = require('mkdirp');
+var util = require('../config/util.js');
+var cloudinary = require('cloudinary');
 
 //
 var UploadPath = path.join(__dirname,'..','public','uploadedfiles');
@@ -20,12 +22,13 @@ var storage = multer.diskStorage({
       var extension = originalname.split(".");
       // cb(null, Date.now() + '.' + extension[extension.length-1]);
       cb(null, Date.now() + '_' + file.originalname);
+      // cb(null,file.originalname);
   }
 });
 var upload = multer({storage : storage});
 //
 
-//Each different photo gallery [Activity, Study, Seminar]
+//Each different photo gallery [Activity, Study, Seminar, Work]
 
 //If connects to photo page - redirect to Activity page
 router.get('/', function(req, res, next) {
@@ -38,14 +41,11 @@ router.get('/Activity',function(req,res){
   var page = req.query.page;
   if(page === undefined) {page = 1;}
   page = parseInt(page);
-  //   //var page = Math.max(1,req.query.page);
-  //   var page = ((req.query.page)===undefined)?1:req.query.page;
-  //   //if(page===undefined) {page = 1;}
   PhotoMods.PhotoMod_Activity.count({}, function(err, count) {
     if(err) return res.json({success:false, message:err});
     var skip = (page-1)*limit;
     var maxPageNum = Math.ceil(count/limit);
-    PhotoMods.PhotoMod_Activity.find({}).sort('-createdAt').skip(skip).limit(limit).exec(function(err, photos) {
+    PhotoMods.PhotoMod_Activity.find({}).populate("author").sort('-createdAt').skip(skip).limit(limit).exec(function(err, photos) {
       if(err) return res.json({success:false, message:err});
       res.render("../views/PhotoGallery/photo",{
   			photodata: photos,
@@ -59,7 +59,7 @@ router.get('/Activity',function(req,res){
    });
  });
 
- router.get('/Activity/new',function(req,res){
+ router.get('/Activity/new',util.isLoggedin,function(req,res){
  	res.render("../views/PhotoGallery/photo_new",{
  			title: 'Photo_MT_Activity',
  			main_menu: 'MT & 활동',
@@ -68,13 +68,13 @@ router.get('/Activity',function(req,res){
  		});
  });
 
-router.post('/Activity',upload.single('photo'),function(req,res){
-console.log("post/activity");
-  if(req.file !== undefined){
-    req.body.filename = req.file.filename;
-    req.body.originalfilename = req.file.originalname;
+router.post('/Activity',util.isLoggedin,upload.fields([{name:'file'},{name:'files'}]),function(req,res){
+  req.body.author = req.user._id;
+  if(req.body.images) {req.body.images = JSON.parse(req.body.images);}
+  if(req.files.file !== undefined) {
+    req.body.filePath = req.files.file[0].path;
+    req.body.fileOriginalname = req.files.file[0].originalname;
   }
-
 	PhotoMods.PhotoMod_Activity.create(req.body,function(err,photos){
 		if(err) return res.json({success:false, message:err});
 		res.redirect('/photo/Activity');
@@ -82,7 +82,7 @@ console.log("post/activity");
 });
 
 router.get('/Activity/:id', function(req, res){
-  PhotoMods.PhotoMod_Activity.findOne({_id:req.params.id}, function(err, photos){
+  PhotoMods.PhotoMod_Activity.findOne({_id:req.params.id}).populate(['author','comments.author']).exec(function(err, photos){
     if(err) return res.json(err);
     res.render('PhotoGallery/photo_view', {
       photodata:photos,
@@ -94,9 +94,10 @@ router.get('/Activity/:id', function(req, res){
   });
 });
 
-router.get('/Activity/:id/edit', function(req,res){
+router.get('/Activity/:id/edit', util.isLoggedin, function(req,res){
   PhotoMods.PhotoMod_Activity.findById(req.params.id, function(err, photos){
     if(err) return res.json({success:false, message:err});
+    if(!req.user._id.equals(photos.author)) return res.json({success:false, message:"Unauthrized Attempt"});
     res.render('PhotoGallery/photo_edit', {
       photodata: photos,
       title: 'Photo_MT_Activity',
@@ -107,28 +108,49 @@ router.get('/Activity/:id/edit', function(req,res){
   });
 });
 
-router.put('/Activity/:id',upload.single('photo'),function(req,res){
+
+router.put('/Activity/:id', util.isLoggedin, upload.fields([{name:'file'},{name:'files'}]),function(req,res){
   if(req.file !== undefined){
     req.body.filename = req.file.filename;
   }
 	req.body.updatedAt=Date.now();
-	PhotoMods.PhotoMod_Activity.findByIdAndUpdate(req.params.id,req.body,function(err,photos){
+	PhotoMods.PhotoMod_Activity.findByIdAndUpdate({_id:req.params.id, author:req.user._id},req.body,function(err,photos){
 		if(err) return res.json({success:false, message:err});
+    if(!photos) return res.json({success:false, message:"No data found to update"});
 		res.redirect(req.params.id + '?page=' + req.query.page);
 	});
 });
 
-router.delete('/Activity/:id', function(req,res){
-	PhotoMods.PhotoMod_Activity.findByIdAndRemove(req.params.id, function(err, photos){
+router.delete('/Activity/:id', util.isLoggedin, function(req,res){
+	PhotoMods.PhotoMod_Activity.findByIdAndRemove({_id:req.params.id, author:req.user._id}, function(err, photos){
 		if(err) {
       return res.json({success:false, message:err});
     }
+    if(!photos) return res.json({success:false, message:"No data found to remove"});
     if(photos.filename !== undefined){
       fs.unlink(path.join(__dirname,'..','public','uploadedfiles',photos.filename));
     }
 		res.redirect('/photo/Activity');
 	});
 });
+
+////comment
+router.post('/Activity/:id/comments',function(req,res){
+	var newComment = req.body.comment;
+	newComment.author = req.user._id;
+	PhotoMods.PhotoMod_Activity.update({_id:req.params.id},{$push:{comments:newComment}},function(err,post){
+		if(err) return res.json({success:false, message:err});
+		res.redirect('/photo/Activity/'+req.params.id);
+	});
+});
+
+router.delete('/Activity/:id/comments/:commentId',function(req,res){
+		PhotoMods.PhotoMod_Activity.update({_id:req.params.id},{$pull:{comments:{_id:req.params.commentId}}},function(err,post){
+		if(err) return res.json({success:false, message:err});
+		res.redirect('/photo/Activity/'+req.params.id);
+	});
+});//destroy
+
 //---Photo_MT_Activity -->
 
 //<-- Photo_Study ---
@@ -164,14 +186,14 @@ router.get('/Study',function(req,res){
  		});
  });
 
-router.post('/Study',upload.single('photo'),function(req,res){
+
+router.post('/Study',upload.fields([{name:'file'},{name:'files'}]),function(req,res){
   if(req.file !== undefined){
     req.body.filename = req.file.filename;
   }
 	PhotoMods.PhotoMod_Study.create(req.body,function(err,photos){
 		if(err) return res.json({success:false, message:err});
-		res.redirect('/photo/Study');
-	});
+		res.redirect('/photo/Study');	});
 });
 
 router.get('/Study/:id', function(req, res){
@@ -200,7 +222,8 @@ router.get('/Study/:id/edit', function(req,res){
   });
 });
 
-router.put('/Study/:id',upload.single('photo'),function(req,res){
+
+router.put('/Study/:id',upload.fields([{name:'file'},{name:'files'}]),function(req,res){
   if(req.file !== undefined){
     req.body.filename = req.file.filename;
   }
@@ -257,7 +280,8 @@ router.get('/Seminar',function(req,res){
  		});
  });
 
-router.post('/Seminar',upload.single('photo'),function(req,res){
+
+router.post('/Seminar',upload.fields([{name:'file'},{name:'files'}]),function(req,res){
   if(req.file !== undefined){
     req.body.filename = req.file.filename;
   }
@@ -293,7 +317,8 @@ router.get('/Seminar/:id/edit', function(req,res){
   });
 });
 
-router.put('/Seminar/:id',upload.single('photo'),function(req,res){
+
+router.put('/Seminar/:id',upload.fields([{name:'file'},{name:'files'}]),function(req,res){
   if(req.file !== undefined){
     req.body.filename = req.file.filename;
   }
@@ -353,8 +378,8 @@ router.get('/Work',function(req,res){
  		});
  });
 
-router.post('/Work',upload.single('photo'),function(req,res){
-console.log("post/Work");
+
+router.post('/Work',upload.fields([{name:'file'},{name:'files'}]),function(req,res){
   if(req.file !== undefined){
     req.body.filename = req.file.filename;
     req.body.originalfilename = req.file.originalname;
@@ -392,7 +417,8 @@ router.get('/Work/:id/edit', function(req,res){
   });
 });
 
-router.put('/Work/:id',upload.single('photo'),function(req,res){
+
+router.put('/Work/:id',upload.fields([{name:'file'},{name:'files'}]),function(req,res){
   if(req.file !== undefined){
     req.body.filename = req.file.filename;
   }
@@ -417,10 +443,15 @@ router.delete('/Work/:id', function(req,res){
 //---Photo_Work-->
 
 
-router.post('/imageupload', upload.single('image'), function(req,res){
-    // console.log(path.join('..','public','uploadedfiles',req.file.filename));
-    req.file.url = path.join('..','..','uploadedfiles',req.file.filename);
-  res.send(req.file);
+//<---Image Upload---
+router.post('/imageupload', upload.single('image'), function(req,res, next){
+    cloudinary.uploader.upload(req.file.path, function(result) {
+      if(result && req.file.path !== undefined){
+        fs.unlink(path.join(req.file.path));
+        console.log("LocalTemporary Image Deleted");
+      }
+      res.send(result);
+     }, {use_filename: true, unique_filename: true });
 });
-
+//---Image Upload--->
 module.exports = router;
